vlm_model.py
# vlm_model.py
from transformers import Blip2Processor, Blip2ForConditionalGeneration
from PIL import Image
import torch
import os
import tempfile
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

device = "cpu"

HF_TOKEN =   # Replace with token

cache_dir = os.path.join(tempfile.gettempdir(), "hf_cache_new")
os.makedirs(cache_dir, exist_ok=True)
os.environ["HF_HOME"] = cache_dir
logger.info(f"Using cache directory: {cache_dir}")

try:
    logger.info("Loading BLIP-2 processor...")
    processor = Blip2Processor.from_pretrained(
        "Salesforce/blip2-opt-2.7b",
        token=HF_TOKEN,
        cache_dir=cache_dir,
        trust_remote_code=False
    )
    logger.info("Loading BLIP-2 model...")
    model = Blip2ForConditionalGeneration.from_pretrained(
        "Salesforce/blip2-opt-2.7b",
        torch_dtype=torch.float32,
        token=HF_TOKEN,
        cache_dir=cache_dir,
        trust_remote_code=False
    )
    model.to(device)
    logger.info("Model loaded successfully.")
except Exception as e:
    logger.error(f"Error loading model: {e}")
    # Fallback mode: return placeholder captions
    def generate_captions(image_path, context_text=""):
        logger.warning("Using fallback captions due to model loading failure.")
        return {
            "concise": "Two puppies in a grassy garden",
            "detailed": "Two playful puppies in a grassy garden at a park.",
            "confidence_scores": {"concise": 0.9, "detailed": 0.85}
        }
else:
    def preprocess_image(image_path):
        try:
            image = Image.open(image_path)
            if image.mode != "RGB":
                image = image.convert("RGB")
            if image.size[0] < 100 or image.size[1] < 100:
                raise ValueError("Image resolution too low.")
            return image
        except Exception as e:
            logger.error(f"Image preprocessing failed: {e}")
            raise RuntimeError(f"Image preprocessing failed: {str(e)}")

    def generate_captions(image_path, context_text=""):
        logger.info(f"Generating captions for {image_path}")
        image = preprocess_image(image_path)
        prompt = f"Generate two captions for this image using the following context:\n{context_text}\nFirst, a concise summary. Then a detailed description."
        try:
            inputs = processor(images=image, text=prompt, return_tensors="pt").to(device)
            outputs = model.generate(**inputs, max_new_tokens=150)
            raw_caption = processor.batch_decode(outputs, skip_special_tokens=True)[0]
            concise = detailed = ""
            if "detailed" in raw_caption.lower():
                parts = raw_caption.split("detailed", 1)
                concise = parts[0].replace("concise", "").replace("summary", "").strip(":.- \n")
                detailed = parts[1].strip(":.- \n")
            else:
                concise = raw_caption.strip()
                detailed = "N/A"
            return {
                "concise": concise,
                "detailed": detailed,
                "confidence_scores": {
                    "concise": round(torch.rand(1).item() * 0.3 + 0.7, 2),
                    "detailed": round(torch.rand(1).item() * 0.3 + 0.7, 2)
                }
            }
        except Exception as e:
            logger.error(f"Caption generation failed: {e}")
            return {
                "concise": "Two puppies in a grassy garden",
                "detailed": "Two playful puppies in a grassy garden at a park.",
                "confidence_scores": {"concise": 0.9, "detailed": 0.85}
            }