requirements.txt
Flask==2.3.2
transformers==4.40.0
torch
opencv-python
Pillow
